feat: add auth-check command to test the tool works (#2)
## Problem

I am trying to setup automated dependency updates for this project. In
order to confidently do that, I need a way to test the tool works when
the dependencies update.

## Solution

Create a simple command `auth-check` that hits a read-only endpoint on
the Google Play API. The idea is that as long as that command runs
successfully, I have asserted the binary compiles, the google cloud sdk
is working, and the auth library is working. This doesn't fully end to
end test everything (like uploading a new APK to the account) but this
is a simple solution to feel *mostly* confident. As long as we make
incremental releases, at least people can pin a release version if they
encounter unstable versions.